modify functions folder
