+++ functions/utils_numpy.py
@@ -1,0 +1,48 @@
+import numpy as np
+
+def quaternions_to_rotation_matrices(quaternions):
+    
+    # initialize
+    K = quaternions.shape[0]
+    R = np.zeros((K, 3, 3))
+
+    # A unit quaternion is q = w + xi + yj + zk
+    xx = quaternions[:, 1]**2
+    yy = quaternions[:, 2]**2
+    zz = quaternions[:, 3]**2
+    ww = quaternions[:, 0]**2
+    n = np.expand_dims((ww + xx + yy + zz), axis=-1)
+    s = np.zeros((K, 1))
+    s[n != 0] = 2 / n[n != 0]
+
+    xy = s[:, 0] * quaternions[:, 1] * quaternions[:, 2]
+    xz = s[:, 0] * quaternions[:, 1] * quaternions[:, 3]
+    yz = s[:, 0] * quaternions[:, 2] * quaternions[:, 3]
+    xw = s[:, 0] * quaternions[:, 1] * quaternions[:, 0]
+    yw = s[:, 0] * quaternions[:, 2] * quaternions[:, 0]
+    zw = s[:, 0] * quaternions[:, 3] * quaternions[:, 0]
+
+    xx = s[:, 0] * xx
+    yy = s[:, 0] * yy
+    zz = s[:, 0] * zz
+
+    idxs = range(K)
+    R[idxs, 0, 0] = 1 - yy - zz
+    R[idxs, 0, 1] = xy - zw
+    R[idxs, 0, 2] = xz + yw
+
+    R[idxs, 1, 0] = xy + zw
+    R[idxs, 1, 1] = 1 - xx - zz
+    R[idxs, 1, 2] = yz - xw
+
+    R[idxs, 2, 0] = xz - yw
+    R[idxs, 2, 1] = yz + xw
+    R[idxs, 2, 2] = 1 - xx - yy
+
+    return R
+
+def define_SE3(R, p):
+    SE3 = np.identity(4)
+    SE3[0:3, 0:3] = R
+    SE3[0:3, 3] = p
+    return SE3
